# Conflicts:
#	Assets/Core/Pooling/PoolingType.cs
#	Assets/font/DOSSaemmul SDF.asset

+++ Assets/DAZB/01.Scripts/Player/PlayerHealth.cs
@@ -18,7 +18,7 @@ public class PlayerHealth : MonoBehaviour, IDamageable
     }
 
     private void Start() {
-        maxHealth = owner.stat.maxHealth.GetValue();
+        maxHealth = (int)owner.stat.maxHealth.GetValue();
         currentHealth = maxHealth;
     }
